Move test class to package `impl`
